+++ scripts/docs/sphinx-build.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 # Work under Documentation directory
-set -e
+set -ex
 
+readonly MYDIR=$(dirname $(realpath $0))
+readonly DOCDIR=$(realpath ${MYDIR}/../../Documentation/)
+
+pushd ${DOCDIR}
 sphinx-build -M html . build
+popd