+++ scripts/scripts-install.sh
@@ -31,6 +31,8 @@ scripts_install()
 	ln -s ${SCRIPTS_DIR}/kcompile.sh /usr/bin/kcompile
 	ln -s ${SCRIPTS_DIR}/qemu-compile.sh /usr/bin/qemu-compile
 	ln -s ${SCRIPTS_DIR}/qemu-vm.sh /usr/bin/qemu-vm
+	ln -s ${SCRIPTS_DIR}/docs/sphinx-serve.sh /usr/bin/sphinx-serve-tl
+	ln -s ${SCRIPTS_DIR}/docs/sphinx-serve.service /usr/lib/systemd/system/sphinx-serve-tl.service
 
 	# Other directory
 	ln -s ${SCRIPTS_DIR}/../ai/pytorch/build/compile /usr/bin/pytorch-compile
@@ -56,7 +58,9 @@ scripts_uninstall()
 		/usr/bin/kcompile \
 		/usr/bin/pytorch-compile \
 		/usr/bin/qemu-compile \
-		/usr/bin/qemu-vm
+		/usr/bin/qemu-vm \
+		/usr/bin/sphinx-serve-tl \
+		/usr/lib/systemd/system/sphinx-serve-tl.service
 }
 
 scripts_set_env()